examples/openai-python/requirements.txt
claiv-memory>=0.6.4
openai>=1.0.0
python-dotenv>=1.0.0
